requirements.txt
Django==4.0.4
psycopg2-binary==2.9.5 # for PostgreSQL. Replace with your db manager
python-dotenv==0.21.0
whitenoise==6.2.0
gunicorn==20.1.0
django-heroku==0.3.1
dj-database-url==1.2.0
djangorestframework==3.14.0
Markdown==3.4.1
django-filter==22.1

